Hook Write to send headers
This fixes GH-14 and the details are there.

If Write is called before WriteHeader, then the headers are never
properly sent down. Although the net/http docs state that Write should
trigger WriteHeader, this happens after httpsnoop so the httpsnoop
callback is never called. We have to manually do this in our Write
callback.